Update Bridge_Jeedom_InfluxDB.py
I use the global push url so any update in jeedom is pushed through the bridge. This is a great feature but there is a lot of different data type to handle.
Bridge is updated to handle string or no value.
Also changed the url decoding to handle special characters.
Thanks for the script.